user, region, rental 3개 흐름하고 코드 정리 및 api 테스트 진행했습니다

+++ src/main/java/com/example/moyeorak/controller/NoticeController.java
@@ -30,7 +30,7 @@ public class NoticeController {
     @PostMapping
     public ResponseEntity<NoticeResponse> createNotice(@RequestBody @Valid NoticeRequest request,
                                                        @AuthenticationPrincipal CustomUserDetails userDetails) {
-        User author = userRepository.findById(Math.toIntExact(userDetails.getUserId()))
+        User author = userRepository.findById((long) Math.toIntExact(userDetails.getUserId()))
                 .orElseThrow(() -> new IllegalArgumentException("작성자 정보를 찾을 수 없습니다."));
 
         NoticeDto created = noticeService.create(author, request);
